back to 16-bit floats

+++ glove.py
@@ -8,6 +8,6 @@ def read(filename):
             first_space_index = line.index(' ')
             word = line[:first_space_index]
             values = line[first_space_index + 1:]
-            vector = np.fromstring(values, sep=' ', dtype=np.float32)
+            vector = np.fromstring(values, sep=' ', dtype=np.float16)
             word_to_vec[word] = vector
     return word_to_vec
